move middlewares to util
